+++ src/Traits/HasPermissionPlus.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Laraditz\PermissionPlus\Traits;
+
+trait HasPermissionPlus
+{
+    public function canAccessPermissionPlus(): bool
+    {
+        return app()->isLocal() ? true : false;
+    }
+}